requirements.txt
google-adk >= 1.22.1
google-adk[a2a]
google-adk[extensions]
python-dotenv >= 1.1.1
litellm >= 1.80.15
uvicorn >= 0.40.0
python-a2a >= 0.5.10
PyPDF2 >= 3.0.1
sentence-transformers >= 5.2.3
qdrant-client >= 1.16.2
nltk >= 3.9.2
crewai >= 1.6.1
crewai[tools]
gradio >= 6.8.0
pytest
httpx >= 0.28.1
deepeval >= 3.8.9

pytest >= 9.0.2